microplastics/requirements.txt
asgiref==3.3.1
backcall==0.2.0
colorama==0.4.4
decorator==5.0.7
Django==3.1.7
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
